update to latest shiny stuff
* readable-stream@3
* remove sauce test garbage
* new tape & bl for testing
* no Buffer constructor in test
* travis update (is it connected?)
* remove license year
