+++ .travis.yml
@@ -1,7 +1,8 @@
 language: node_js
 node_js:
-- '0.10'
-- '0.11'
+- '6'
+- '8'
+- '10'
 branches:
   only:
   - master
@@ -9,8 +10,3 @@ notifications:
   email:
   - rod@vagg.org
 script: npm test
-env:
-  global:
-  - SAUCE_USER=through2-sauce
-  - secure: A8KyygIlklZOcTwTGhZHpkOu7Cn4tI1QQasg75CTSjudczGzkuh1pG2RDIkWqGcrWv9o5CjI6fFYYx0JvQu8o5PO+ZAtV1Z0AlnM965vE8kAx0XEC7giAV323bibfnHwyQu74vIhbOfdes7E4/cYerV2kAjK0T375D4DYEpDuvw=
-
