+++ package.json
@@ -4,8 +4,7 @@
   "description": "A tiny wrapper around Node streams2 Transform to avoid explicit subclassing noise",
   "main": "through2.js",
   "scripts": {
-    "test": "node test/test.js | faucet",
-    "test-local": "brtapsauce-local test/basic-test.js"
+    "test": "node test/test.js | faucet"
   },
   "repository": {
     "type": "git",
@@ -20,13 +19,13 @@
   "author": "Rod Vagg <r@va.gg> (https://github.com/rvagg)",
   "license": "MIT",
   "dependencies": {
-    "readable-stream": "^2.1.5",
+    "readable-stream": "^3.0.1",
     "xtend": "~4.0.1"
   },
   "devDependencies": {
-    "bl": "~1.1.2",
+    "bl": "~2.0.1",
     "faucet": "0.0.1",
-    "stream-spigot": "~3.0.5",
-    "tape": "~4.6.2"
+    "stream-spigot": "~3.0.6",
+    "tape": "~4.9.1"
   }
 }
